svn: Add a tsh_blame_baton wrapper
In svn 1.12, blame5 has been deprecated, the callback of blame6 omits
some arguments that was available in blame5's callback.  Callers are
expected to take that from output arguments of blame6 instead.
Thus, we would need to a wrapper to provides that information to our
callbacks.

Strictly speaking, those informations aren't used as of it's.
However, it's implementation detail.  Let's create a wrapper to make
them on pair with each others.